+++ Demo/Demo/ARAppDelegate.m
@@ -21,8 +21,8 @@
     ARCollectionViewMasonryLayout *layout = [[ARCollectionViewMasonryLayout alloc] initWithDirection:ARCollectionViewMasonryLayoutDirectionVertical];
     ARCollectionViewController *viewController = [[ARCollectionViewController alloc] initWithCollectionViewLayout:layout];
     viewController.colorCount = 100;
-    viewController.headerHeight = 5;
-    viewController.footerHeight = 2;
+    viewController.headerSize = CGSizeMake(10, 5);
+    viewController.footerSize = CGSizeMake(10, 2);
     self.window.rootViewController = viewController;
     
     [self.window makeKeyAndVisible];